compositor/native: Enable frame synchronization for surface actors
Find a surface actor that meets the criteria to drive the refresh
rate. If a new surface actor is found, add a reference to it and
request frame synchronization to be enabled for the
MetaRendererViewNative.

Whenever the surface actor schedules a repaint, and in case frame
synchronization is enabled for the MetaRendererViewNative, schdule
the repaint to occur immediately. This effectively makes the surface
actor's frame rate drive the refresh rate of the monitor.

If the actor is frozen or destroyed, it is no longer expected to
schedule repaints and should stop driving the refresh rate.

When there is no surface actor to drive the refresh rate, request
frame synchronization to be disabled for the MetaRendererViewNative.