renderer-view/native: Allow requesting frame synchronization
Frame synchronization is enabled as long as it's requested to be
enabled and is applicable to be enabled. It is applicable to be
enabled if VRR is supported and not disallowed for the output.

When frame synchronization is enabled, it enables the the variable
mode of ClutterFrameClock and enables VRR for the output.

Changes in the frame synchronization mode are applied asynchronously,
before the next frame is drawn.