clutter/frame-clock: Add a mode for variable scheduling
A new variable scheduling mode is introduced which allows lower
priority updates to be schduled on a timeout which represents
a lower refresh rate. while allowing higher priority updates
that are scheduled to occur "now" to occur immediately.

This mode will be used by following commits to implement
synchronization of page flips to the update rate of actors
that drive the frame clock, while deferring other updates (such
as cursor updates) to a lower refresh rate.

This is needed to avoid abrupt changes in the variable refresh
rate due to cursor movement, which in turn results in brightness
flicker.