backends/native: Add udev rule to disable VRR for i915
This adds a udev rule that disables support for VRR with i915
for now.

The i915 DRM driver requires DRM_MODE_ATOMIC_ALLOW_MODESET to be set
on the atomic commit when updating the "VRR_ENABLED" KMS property,
which effectively means that it can perform a full modeset when
enabling or disabling VRR dynamically.

Since a full modeset can result in blanking of the monitor and bad
user experience, it is avoided by default.

While this limitation has been lifted with recent kernel changes,
most released kernel versions are still affected.