Reduce entropy of .symbols file
The SONAME changed in version 43~beta, so there's no point in keeping
track of whether symbols are older than that version.

Similarly, if a symbol was introduced in a prerelease, we want
dependent packages to depend on the final release, so there's little
point in finer-grained tracking of which specific prerelease each
symbol came from.