d/patches, d/rules: Stop trying to force gles2 driver on armel, armhf
This patch was rejected by upstream, and according to comments on the
upstream merge request, it is no longer effective either, because
mutter's driver selection is now done in a higher layer.
Users of systems with a working gles2 driver but a broken or slow gl3
driver can work around this by setting environment variable
CLUTTER_DRIVER=gles2.

See #1050014 for more details of what would be involved in making this
driver choice happen automatically in an upstreamable way.